% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearTutorial.R
\name{clearTutorial}
\alias{clearTutorial}
\title{Clearing a tutorial for MicroWithR.}
\usage{
clearTutorial(module, lesson, practice = FALSE, ...)
}
\arguments{
\item{module}{is a numerical value corresponding to the module.
0 = Introduction
1 = Consumer Equilibrium
2 = Producer Equilibrium
3 = Market Equilibrium}

\item{lesson}{is a numerical value corresponding to the lesson within a module.}

\item{practice}{is a logical value indicating whether to clear the main tutorial or practice.}
}
\description{
This function clears previous recordings of a tutorial.
}
