% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRandomPresenterApp.R
\name{runRandomPresenterApp}
\alias{runRandomPresenterApp}
\title{Selecting a random student to present a tutorial.}
\usage{
runRandomPresenterApp()
}
\description{
This function runs the \code{randomPresenterApp} from \code{shiny-examples} to select
a random presenter.
}
