% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTutorial.R
\name{runTutorial}
\alias{runTutorial}
\title{Running a tutorial for MicroWithR.}
\usage{
runTutorial(module, lesson, practice = FALSE, ...)
}
\arguments{
\item{module}{is a numerical value corresponding to the module.
0 = Introduction
1 = Consumer Equilibrium
2 = Producer Equilibrium
3 = Market Equilibrium}

\item{lesson}{is a numerical value corresponding to the lesson within a module.}

\item{practice}{is a logical value indicating whether to clear the main tutorial or practice.}

\item{data}{is the universe of all variables considered, which may include
both controls and excluded variables (e.g. variables eliminated by a model
selection algorithm).}
}
\description{
This function runs a \code{MicroWithR} tutorial.
}
